Add benchmarks for message processing for TCP and Datagram connections
The reason we can't simply use ListenUDP() and Dial() in the datagram
benchmark is that we wouldn't be able to ensure that all messages would
arrive on the server, even for a udp connection to localhost.
By using the fakePacketConn we ensure that all messages are delivered
making the datagram benchmark reliable.