Fix up parseUpToLen(). (#46)
The parseUpToLen function would not handle the max chars limit.
Found when using a syslog client which used > max chars for the
APP-NAME. I changed the syslog client to use max chars and it still
didn't work. There was an off by 1 error. One of the tests even
gave a max chars string and expected an error - changed that test
to try with max+1 chars.